requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
numpy>=1.25.0
matplotlib>=3.8.0
scipy>=1.11.0
sqlalchemy>=2.0.0
tabulate>=0.9.0
openai>=1.30.0
tiktoken>=0.5.2
python-dotenv>=1.0.1
